msp430/x5: McuSleep/msp430PowerBits should have LPM2 includ SCG0 for FLL disable.
The TI documentation for the 5438 (all x5 family?) says the FLL is disabled if
SCG0 or SCG1 are set (SR) and yes that is true.   If either bit is set then the
FLL isn't driving the DCO (either because the FLL integrator is turned off or the
DCO itself is turned off).   In LPM2, SCG1 is set which disables the DCO generator
(and the FLL).  But if SCG0 is clear it looks like the frequency integrator is
still running.   Which doesn't buy us anything.   The DCO is off so what is the
point.   So it seems to make sense to kick both SCG0 and SCG1.